Change functionality of contact info input.
All fields are now editable, and clicking the + button now creates a new blank one, rather than having a single form to fill out that adds them. This functionality is much more intuitive, and solves the problem of editing addresses, as well as some of the odd behaviour of the save button.

This resolves #88, resolves #106, and resolves #87.