Update muster DTO to match backend DTO.
Due to the new new DTOs in the backend, the value in the dictionary is expected to be an object containing both the id of the muster status and any remarks you want to put on the muster.  Because you don't use remarks, I just set it to null.

I'm sorry if I fucked up the syntax :D  You have your headphones in and I didn't want to disturb you.